src/requirements.txt
alembic==1.6.5
anyio==3.3.0
APScheduler==3.7.0
argon2-cffi==20.1.0
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bidict==0.21.2
bleach==4.0.0
cached-property==1.5.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
dnspython==2.1.0
entrypoints==0.3
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
Flask-SSE==1.0.0
greenlet==1.1.1
h11==0.12.0
idna==3.2
iexfinance==0.5.0
inflection==0.5.1
ipykernel==5.5.0
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.10.2
jupyterlab==3.1.7
jupyterlab-pygments==0.1.2
jupyterlab-server==2.7.1
Mako==1.1.5
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
mypy-extensions==0.4.3
mysqlclient==2.0.3
nbclassic==0.3.1
nbclient==0.5.4
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.3
nulltype==2.3.1
numpy==1.19.5
packaging==21.0
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
plaid-python==8.0.0
polygon-api-client==0.2.11
prometheus-client==0.11.0
prompt-toolkit==3.0.19
ptyprocess==0.7.0
py-bcrypt==0.4
pycparser==2.20
Pygments==2.10.0
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-editor==1.0.4
python-engineio==4.2.1
python-socketio==5.4.0
pytz==2021.1
pyzmq==22.2.1
redis==3.5.3
requests==2.26.0
requests-unixsocket==0.2.0
Send2Trash==1.8.0
simple-websocket==0.3.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.23
sqlalchemy-orm==1.2.2
sseclient-py==1.7
terminado==0.11.1
testpath==0.5.0
tornado==6.1
traitlets==4.3.3
typing-extensions==3.10.0.0
typing-inspect==0.7.1
tzlocal==2.1
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
websockets==9.1
Werkzeug==2.0.1
wsproto==1.0.0

contextvars~=2.4
pip~=21.1.2
colorama~=0.4.4
dataclasses~=0.8
setuptools~=57.0.0